BUG Numeric only backup codes can be problematic
PHP casts all numeric strings that can be integers into integers when
they are used as the key in an array. This can cause issues with the
backup code generation, however is currently only really affecting the
tests as shown in this commit. The potential for larger issues is there
however, and this should be addressed. Commit is a test only, in a test
driven development type example of an issue to fix.